moved PlayerAvatar into a shared component
