+++ src/features/Shared/PlayerAvatar/PlayerRole/index.js
@@ -1,0 +1,1 @@
+export { PlayerRole } from './PlayerRole'
