+++ src/features/Shared/PlayerAvatar/index.js
@@ -1,0 +1,1 @@
+export { PlayerAvatar } from './PlayerAvatar'
